+++ fantaski-fe/src/config/url.js
@@ -18,3 +18,11 @@ export const ORDERIMAGE_URL =
 export const COURSE_IMG_URL =
   process.env.REACT_APP_COURSE_IMG_URL ||
   "http://localhost:3000/assets/images_course";
+
+export const IMAGE_ROUTER_URL =
+  process.env.REACT_APP_IMAGE_ROUTER_URL ||
+  "http://localhost:3000/assets/images_route";
+
+export const IMAGE_MEMBER_URL =
+  process.env.REACT_APP_IMAGE_MEMBER_URL ||
+  "http://localhost:3000/assets/images_member";
